Simplify MviView concept
